chore: move awardsComponent into LeftTab
